requirements.txt
django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
celery==5.5.3
redis==6.4.0
minio==7.2.16
boto3==1.40.21
django-storages==1.14.6
drf-spectacular==0.28.0
uvicorn==0.35.0
openai==1.102.0
pytest==8.4.1
pytest-django==4.11.1